# chore: Add Benchmark Scripts and Performance Comparison of LitData vs FFCV for Streaming ImageNet