--- a/benchmarks/litdata/stream_imagenet.py
+++ b/benchmarks/litdata/stream_imagenet.py
@@ -99,6 +99,7 @@ def main():
         ),
         batch_size=args.batch_size,
         num_workers=args.num_workers,
+        drop_last=True,
     )
 
     print(