## What does this PR do?

- [x] Introduces benchmark scripts for streaming ImageNet using LitData and FFCV  
- [x] Provides a performance comparison between the two frameworks

### Benchmarks for LitData vs *FFCV*

Speed to stream Imagenet 1.2M from local disk with ffcv vs LitData:
| Framework | Dataset Mode | Dataset Size @ 256px | Images / sec 1st Epoch (float32) | Images / sec 2nd Epoch (float32) |
|---|---|---|---|---|
| LitData | PIL RAW | 168 GB | 6647 | 6398 | 
| LitData | JPEG 90% | 12 GB | 6553 | 6537 |
| ffcv (os_cache=True) | RAW | 170 GB | 7263 | 6698 |
| ffcv (os_cache=False) | RAW | 170 GB | 7556 | 8169 |
| ffcv(os_cache=True) | JPEG 90% | 20 GB | 7653 | 8051 |
| ffcv(os_cache=False) | JPEG 90% | 20 GB | 8149 | 8607 |

<details><summary>Benchmark Logs</summary>

<p>

**LitData | JPEG 90% | 12 GB**
```python
⚡~/litData/benchmarks aws s3 cp --recursive s3://grid-cloud-litng-ai-03/projects/01j12bk075bs7e771jhmgvz7eg
/datasets/imagenet-1m-litdata/train_256_jpg_90 data/imagenet-1m-litdata/train_256_jpg_90

⚡~/litData/benchmarks python litdata/stream_imagenet.py --input_dir data/imagenet-1m-litdata/train_256_jpg_90 
[INFO] Running streaming benchmark with arguments: Namespace(input_dir='data/imagenet-1m-litdata/train_256_jpg_90', cache_dir='/cache/chunks', dtype='float32', batch_size=256, num_workers=32, drop_last=False, epochs=2, max_cache_size='200GB', use_pil=False, clear_cache=True)
Seed set to 42
[INFO] Clearing cache directory: /cache/chunks
[INFO] Initializing streaming dataset from: data/imagenet-1m-litdata/train_256_jpg_90
[INFO] Starting benchmark for 2 epoch(s) with batch size 256 and 32 workers.
Epoch 1/2: 100%|██████████████████████████████████████████████████████████████████████████████████████| 5005/5005 [03:15<00:00, 25.60it/s]
[RESULT] Epoch 1: Streamed 1281167 samples in 195.51s (6552.99 images/sec)
Epoch 2/2: 100%|██████████████████████████████████████████████████████████████████████████████████████| 5005/5005 [03:15<00:00, 25.54it/s]
[RESULT] Epoch 2: Streamed 1281167 samples in 195.99s (6536.99 images/sec)
[INFO] Clearing cache directory after benchmark: /cache/chunks
[INFO] Finished streaming benchmark.
```

**LitData | RAW | 168 GB**
```python
⚡~/litData/benchmarks aws s3 cp --recursive s3://grid-cloud-litng-ai-03/projects/01j12bk075bs7e771jhmgvz7eg/datasets/imagenet-1m-litdata/train_256_raw_pil data/imagenet-1m-litdata/train_256_raw_pil

⚡~/litData/benchmarks python litdata/stream_imagenet.py --input_dir data/imagenet-1m-litdata/train_256_raw_pil --use_pil 
Seed set to 42
[INFO] Running streaming benchmark with arguments: Namespace(input_dir='data/imagenet-1m-litdata/train_256_raw_pil', cache_dir='/cache/chunks', dtype='float32', batch_size=256, num_workers=32, drop_last=False, epochs=2, max_cache_size='200GB', use_pil=True, clear_cache=True)
[INFO] Clearing cache directory: /cache/chunks
[INFO] Initializing streaming dataset from: data/imagenet-1m-litdata/train_256_raw_pil
[INFO] Starting benchmark for 2 epoch(s) with batch size 256 and 32 workers.
Epoch 1/2: 100%|██████████████████████████████████████████████████████████████████████████████████████| 5005/5005 [03:12<00:00, 25.97it/s]
[RESULT] Epoch 1: Streamed 1281167 samples in 192.74s (6647.00 images/sec)
Epoch 2/2: 100%|██████████████████████████████████████████████████████████████████████████████████████| 5005/5005 [03:20<00:00, 24.99it/s]
[RESULT] Epoch 2: Streamed 1281167 samples in 200.24s (6398.02 images/sec)
[INFO] Clearing cache directory after benchmark: /cache/chunks
[INFO] Finished streaming benchmark.
```

**FFCV | JPEG 90% | 20 GB**
```python
⚡~/litData/benchmarks aws s3 cp s3://grid-cloud-litng-ai-03/projects/01j12bk075bs7e771jhmgvz7eg/datasets/imagenet-1m-ffcv/train_256_100.0_90.ffcv data/imagenet-1m-ffcv/train_256_100.0_90.ffcv

⚡~/litData/benchmarks python ffcv/stream_imagenet.py --cfg.data_path data/imagenet-1m-ffcv/train_256_100.0_90.ffcv --cfg.os_cache True 
┌ Arguments defined───────────────────────────────────────────────┐
│ Parameter       │ Value                                         │
├─────────────────┼───────────────────────────────────────────────┤
│ cfg.data_path   │ data/imagenet-1m-ffcv/train_256_100.0_90.ffcv │
│ cfg.batch_size  │ 256                                           │
│ cfg.num_workers │ 32                                            │
│ cfg.drop_last   │ False                                         │
│ cfg.epochs      │ 2                                             │
│ cfg.order       │ SEQUENTIAL                                    │
│ cfg.os_cache    │ True                                          │
│ cfg.normalize   │ False                                         │
└─────────────────┴───────────────────────────────────────────────┘
Seed set to 42
[INFO] Starting streaming benchmark...
Epoch 1/2: 100%|██████████████████████████████████████████████████████████████████████████████████████| 5005/5005 [02:47<00:00, 29.90it/s]
[RESULT] Epoch 1: Streamed 1281167 samples in 167.40s (7653.53 images/sec)
Epoch 2/2: 100%|██████████████████████████████████████████████████████████████████████████████████████| 5005/5005 [02:39<00:00, 31.45it/s]
[RESULT] Epoch 2: Streamed 1281167 samples in 159.12s (8051.64 images/sec)
[INFO] Finished streaming benchmark.

⚡~/litData/benchmarks python ffcv/stream_imagenet.py --cfg.data_path data/imagenet-1m-ffcv/train_256_100.0_90.ffcv
┌ Arguments defined───────────────────────────────────────────────┐
│ Parameter       │ Value                                         │
├─────────────────┼───────────────────────────────────────────────┤
│ cfg.data_path   │ data/imagenet-1m-ffcv/train_256_100.0_90.ffcv │
│ cfg.batch_size  │ 256                                           │
│ cfg.num_workers │ 32                                            │
│ cfg.drop_last   │ False                                         │
│ cfg.epochs      │ 2                                             │
│ cfg.order       │ SEQUENTIAL                                    │
│ cfg.os_cache    │ False                                         │
│ cfg.normalize   │ False                                         │
└─────────────────┴───────────────────────────────────────────────┘
Seed set to 42
[INFO] Starting streaming benchmark...
Epoch 1/2: 100%|██████████████████████████████████████████████████████████████████████████████████████| 5005/5005 [02:37<00:00, 31.84it/s]
[RESULT] Epoch 1: Streamed 1281167 samples in 157.21s (8149.52 images/sec)
Epoch 2/2: 100%|██████████████████████████████████████████████████████████████████████████████████████| 5005/5005 [02:28<00:00, 33.62it/s]
[RESULT] Epoch 2: Streamed 1281167 samples in 148.85s (8607.16 images/sec)
[INFO] Finished streaming benchmark.
```


**FFCV | RAW | 170GB**
```python
⚡~/litData/benchmarks aws s3 cp s3://grid-cloud-litng-ai-03/projects/01j12bk075bs7e771jhmgvz7eg/datasets/imagenet-1m-ffcv/train_256_0.0_100.ffcv data/imagenet-1m-ffcv/train_256_0.0_100.ffcv

⚡~/litData/benchmarks python ffcv/stream_imagenet.py --cfg.data_path data/imagenet-1m-ffcv/train_256_0.0_100.ffcv --cfg.os_cache True
┌ Arguments defined──────────────────────────────────────────────┐
│ Parameter       │ Value                                        │
├─────────────────┼──────────────────────────────────────────────┤
│ cfg.data_path   │ data/imagenet-1m-ffcv/train_256_0.0_100.ffcv │
│ cfg.batch_size  │ 256                                          │
│ cfg.num_workers │ 32                                           │
│ cfg.drop_last   │ False                                        │
│ cfg.epochs      │ 2                                            │
│ cfg.order       │ SEQUENTIAL                                   │
│ cfg.os_cache    │ True                                         │
│ cfg.normalize   │ False                                        │
└─────────────────┴──────────────────────────────────────────────┘
Seed set to 42
[INFO] Starting streaming benchmark...
Epoch 1/2: 100%|██████████████████████████████████████████████████████████████████████████████████████| 5005/5005 [02:56<00:00, 28.38it/s]
[RESULT] Epoch 1: Streamed 1281167 samples in 176.37s (7263.97 images/sec)
Epoch 2/2: 100%|██████████████████████████████████████████████████████████████████████████████████████| 5005/5005 [03:11<00:00, 26.17it/s]
[RESULT] Epoch 2: Streamed 1281167 samples in 191.26s (6698.55 images/sec)
[INFO] Finished streaming benchmark.

⚡~/litData/benchmarks python ffcv/stream_imagenet.py --cfg.data_path data/imagenet-1m-ffcv/train_256_0.0_100.ffcv
┌ Arguments defined──────────────────────────────────────────────┐
│ Parameter       │ Value                                        │
├─────────────────┼──────────────────────────────────────────────┤
│ cfg.data_path   │ data/imagenet-1m-ffcv/train_256_0.0_100.ffcv │
│ cfg.batch_size  │ 256                                          │
│ cfg.num_workers │ 32                                           │
│ cfg.drop_last   │ False                                        │
│ cfg.epochs      │ 2                                            │
│ cfg.order       │ SEQUENTIAL                                   │
│ cfg.os_cache    │ False                                        │
│ cfg.normalize   │ False                                        │
└─────────────────┴──────────────────────────────────────────────┘
Seed set to 42
[INFO] Starting streaming benchmark...
Epoch 1/2: 100%|██████████████████████████████████████████████████████████████████████████████████████| 5005/5005 [02:49<00:00, 29.52it/s]
[RESULT] Epoch 1: Streamed 1281167 samples in 169.54s (7556.78 images/sec)
Epoch 2/2: 100%|██████████████████████████████████████████████████████████████████████████████████████| 5005/5005 [02:36<00:00, 31.92it/s]
[RESULT] Epoch 2: Streamed 1281167 samples in 156.82s (8169.70 images/sec)
[INFO] Finished streaming benchmark.
```

</p>
</details> 

## PR review

Anyone in the community is free to review the PR once the tests have passed.
If we didn't discuss your PR in GitHub issues there's a high chance it will not be merged.

## Did you have fun?

Make sure you had fun coding 🙃