@bhimrazy 2025-05-13
> Can you add a folder with the files to do the benchmarks ?

Sure @tchaton.